+++ src/config/options.js
@@ -1,23 +1,16 @@
 export default {
-  task_param_prefix: '-',
-  paths: {
-    tasks: 'fbi/',
-    config: 'fbi/config.js'
+  PATHS: {
+    local: {
+      tasks: 'fbi/',
+      config: 'fbi/config.js'
+    },
+    global: {
+      root: 'data',
+      tasks: 'data/tasks',
+      templates: 'data/templates'
+    }
   },
-  data: {
-    root: '../data',
-    tasks: '../data/tasks',
-    templates: '../data/templates'
-  },
-  server: {
-    root: './',
-    host: 'localhost',
-    port: 8888
-  },
-  npm: {
-    alias: 'npm',
-    options: ''
-  },
+  TASK_PARAM_PREFIX: '-',
   TEMPLATE_ADD_IGNORE: ['node_modules', '.DS_Store', '.svn', '.git', 'dst', 'dist'],
   TEMPLATE_INIT_IGNORE: ['node_modules', '.DS_Store', '.svn', '.git', 'dst', 'dist'],
   BACKUP_IGNORE: ['node_modules', '.DS_Store', '.svn', '.git', 'dst', 'dist'],