+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### v3.0.0-alpha.1
+2016.12.12
+- Breaking change:
+    1. global 'require' rewrite
+    1. global 'ctx' injection
+    1. with or without FBI
+- GLobal options change
+
 ### v3.0.0-alpha.0
 2016.12.12
 - init